requirements.txt
aiogram==2.19
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
Babel==2.9.1
bcrypt==3.2.0
bidict==0.21.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
Deprecated==1.2.13
distro==1.7.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
flake8==4.0.1
frozenlist==1.3.0
idna==3.3
iniconfig==1.1.1
jsonschema==3.2.0
mccabe==0.6.1
mongo==0.2.0
multidict==6.0.2
packaging==21.3
paramiko==2.10.1
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
pymongo==4.0.2
PyNaCl==1.5.0
pyparsing==3.0.7
pyrsistent==0.18.1
pytest==7.1.1
pytest-asyncio==0.18.3
python-dotenv==0.19.2
python-engineio==3.14.2
python-socketio==4.6.0
pytz==2021.3
PyYAML==5.4.1
redis==4.1.4
requests==2.27.1
six==1.16.0
texttable==1.6.4
toml==0.10.2
tomli==2.0.1
urllib3==1.26.8
websocket-client==0.59.0
websockets==10.2
wrapt==1.13.3
yarl==1.7.2
